Rename algorithm_sort to main, Move all sorts to sorting_class
