+++ word_processing.rb
@@ -1,7 +1,8 @@
+# Class does not have any function yet.
 class WordProcessing
   # A method to check for capitalization
   def capitalized?(word)
     word == word.capitalize
   end
-  
+
 end
